Modifed the file of Limit operator
